table changes

+++ commands/processor.commands.go
@@ -18,13 +18,19 @@ func (t *TgCommandProcessor) processCommands(ctx context.Context, message *tgbot
 	switch command {
 	case "start":
 		_, _ = SendMsg(t.bot, chatID, "Hello there! Just send me full URL that you want to track.")
-		err = t.botFlowRepository.Save(ctx, chatID, domain.URLRequsted)
-		t.contentRepository.Create(ctx, chatID)
+		err = t.botFlowRepository.Save(ctx, chatID, domain.URLRequsted, nil)
 
 	case "stop":
 		_, _ = SendMsg(t.bot, chatID, "Bye bye!")
-		err = t.botFlowRepository.Delete(ctx, chatID)
-		t.contentRepository.Delete(ctx, chatID)
+		flowStep, savedUrl, _ := t.botFlowRepository.FindByChatID(ctx, chatID)
+
+		if flowStep != domain.NotStarted {
+			err = t.botFlowRepository.Delete(ctx, chatID)
+
+			if savedUrl != nil {
+				t.contentRepository.Delete(ctx, chatID, *savedUrl)
+			}
+		}
 
 	default:
 		_, err = SendMsg(t.bot, chatID, "Sorry, I don't understand that command. Please use /start to get started.")
